requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Database & Storage (Cloud)
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
geoalchemy2==0.14.2
pandas==2.1.3
pyarrow==14.0.1

# Cloud AI Services
openai>=1.6.1
cohere==4.37
groq==0.4.1

# Vector Database (Pinecone)
pinecone-client==2.2.4



# Redis Cloud
redis==5.0.1

# Missing Dependencies
pydantic-settings==2.0.3
asyncpg==0.29.0
aiofiles==23.2.1
aiohttp==3.9.1

# Geospatial & Ocean Data
xarray==2023.11.0
netcdf4==1.6.5
cartopy==0.22.0
folium==0.15.0
geopandas==0.14.1
shapely==2.0.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
bokeh==3.3.0

# API & Web
requests==2.31.0
httpx==0.25.2
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.3.2



# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0